+++ sandpack-react/src/contexts/utils/useClient.ts
@@ -217,8 +217,14 @@ export const useClient: UseClient = ({ options, customSetup }, filesState) => {
   const runSandpack = useCallback(async (): Promise<void> => {
     await Promise.all(
       Object.keys(preregisteredIframes.current).map(async (clientId) => {
+        // There's already a client if the same id, so we should destroy it
+        if (clients.current[clientId]) {
+          clients.current[clientId].destroy();
+        }
+        
         const { iframe, clientPropsOverride = {} } =
           preregisteredIframes.current[clientId];
+          
         clients.current[clientId] = await createClient(
           iframe,
           clientId,